Make status register private, add getter/setter/clear
Make the status register private and use @property to control get/set of
the status register. Bit 5 seems to be required to 1, perhaps we can
control illegally or accidently setting it to a value if we decide to
set the whole status register in 1 operation.
Added functions to set and clear the carry flag, as a template for the
rest of the flags?