Merging with bobomb-dev & adding reference code
Found a transistor-level C implementation of the 6502 and a unittest for
each instruction. Will use its output to extract opcode cycle timings &
other valuable data for opcode implementation.

Checked-in code is licensed under an open-source but non-copyleft
license. Please see each source file and the COPYING file in the
directory for more info. (THIS CODE IS NOT USED IN dNES ITSELF)